Moved Feature Vector and Design Matrix to CommonUtilities Project
