+++ SignalClassifierPrototype/.vs/SignalClassifierPrototype/project-colors.json
@@ -10,7 +10,12 @@
       "ProjectGuid": "a2fe74e1-b743-11d0-ae1a-00a0c90fffc3",
       "DisplayName": "Miscellaneous Files",
       "ColorIndex": -1
+    },
+    "0b3aab92-77da-4d35-8585-318b4458c673": {
+      "ProjectGuid": "0b3aab92-77da-4d35-8585-318b4458c673",
+      "DisplayName": "CommonUtilities",
+      "ColorIndex": 1
     }
   },
-  "NextColorIndex": 1
+  "NextColorIndex": 2
 }
