+++ SignalClassifierPrototype/CommonUtilities/__init__.py
@@ -1,0 +1,9 @@
+"""
+Repository:     Buell-Senior-Thesis
+Solution:       SignalClassifierPrototype
+Project:        FeatureCollection  
+File:           __init__.py
+ 
+Author:         Landon Buell
+Date:           December 2021
+"""
